+++ apps/cli/src/index.ts
@@ -13,13 +13,17 @@ import VariableCommand from './commands/variable.command'
 import { version } from '../package.json'
 import ConfigCommand from './commands/config.command'
 import IntegrationCommand from './commands/integration.command'
+import LoginCommand from './commands/login.command'
 
 const program = new Command()
 
 program.version(version, '-V, --version', 'Output the current version')
-program.option('--profile <string>', 'The profile to use')
-program.option('--api-key <string>', 'The API key to use')
-program.option('--base-url <string>', 'The base URL to use')
+program.option('-p, --profile <string>', 'The profile to use')
+program.option(
+  '-k, --token <string>',
+  'Token used to authenticate your requests'
+)
+program.option('-u, --base-url <string>', 'The base URL to use')
 
 const COMMANDS: BaseCommand[] = [
   new RunCommand(),
@@ -32,7 +36,8 @@ const COMMANDS: BaseCommand[] = [
   new ScanCommand(),
   new VariableCommand(),
   new ConfigCommand(),
-  new IntegrationCommand()
+  new IntegrationCommand(),
+  new LoginCommand()
 ]
 
 COMMANDS.forEach((command) => {