+++ apps/platform/src/components/dashboard/project/createProjectDialogue/create-project-name.tsx
@@ -10,15 +10,15 @@ export default function CreateProjectName({
   onChange
 }: CreateProjectNameProps): React.JSX.Element {
   return (
-    <div className="flex h-[2.25rem] w-full items-center gap-7">
+    <div className="flex h-9 w-full items-center gap-7">
       <Label
-        className="font-geist h-[1.25rem] w-[4.813rem] gap-[0.25rem] text-left text-[0.875rem] font-[500] "
+        className="font-geist h-5 w-[4.813rem] gap-1 text-left text-[0.875rem] font-medium "
         htmlFor="name"
       >
         Name
       </Label>
       <Input
-        className="col-span-3 h-[2.25rem] w-[20rem] "
+        className="col-span-3 h-9 w-[20rem] "
         id="name"
         onChange={(e) => {
           onChange(e.target.value)