+++ apps/platform/src/components/shared/sidebar/index.tsx
@@ -1,6 +1,7 @@
 'use client'
 import React from 'react'
 import { useAtomValue } from 'jotai'
+import Link from 'next/link'
 import {
   CardSVG,
   DashboardSVG,
@@ -10,8 +11,8 @@ import {
   SettingsSVG,
   TeamSVG
 } from '@public/svg/shared'
-import SidebarTab from './sidebarTab'
 import TierLimit from './tierLimit'
+import SidebarTab from './sidebarTab'
 import { Combobox } from '@/components/ui/combobox'
 import { selectedWorkspaceAtom } from '@/store'
 import { VERSION_BADGE } from '@/constants/sidebar'
@@ -59,7 +60,7 @@ function Sidebar(): JSX.Element {
   ]
 
   return (
-    <aside className="ml-4 h-screen w-[18rem] min-w-[16rem] flex-shrink-0">
+    <aside className="ml-4 h-screen w-[18rem] min-w-[16rem] shrink-0">
       <div className="flex h-full flex-col gap-2">
         {/* Sticky Header */}
         <div className="sticky top-0 z-50 bg-[#0A0A0A] pb-4">
@@ -67,7 +68,7 @@ function Sidebar(): JSX.Element {
             <div className=" flex gap-2 text-xl">
               <KeyshadeLogoSVG /> Keyshade
             </div>
-            <div className="rounded bg-white/10 px-2 py-[0.12rem] text-xs font-bold">
+            <div className="rounded-sm bg-white/10 px-2 py-[0.12rem] text-xs font-bold">
               {VERSION_BADGE}
             </div>
           </div>
@@ -77,9 +78,7 @@ function Sidebar(): JSX.Element {
         </div>
 
         {/* Scrollable Menu Items */}
-        <div
-          className="scrollbar-thin scrollbar-thumb-white/20 scrollbar-track-transparent min-h-0 flex-1 overflow-y-auto overscroll-contain"
-        >
+        <div className="scrollbar-thin scrollbar-thumb-white/20 scrollbar-track-transparent min-h-0 flex-1 overflow-y-auto overscroll-contain">
           {sidebarTabData.map((tabData) => {
             return (
               <SidebarTab
@@ -93,8 +92,16 @@ function Sidebar(): JSX.Element {
           })}
         </div>
 
+        {/* Contact Us */}
+        <Link
+          className="relative flex w-full gap-x-3 rounded-md p-2.5 capitalize transition-colors hover:text-white/60"
+          href="mailto:support@keyshade.io?subject=Query"
+        >
+          <div>Contact Us</div>
+        </Link>
+
         {/* Sticky Footer */}
-        <div className="sticky bottom-0 z-50 bg-[#0A0A0A] pt-4 mb-5">
+        <div className="sticky bottom-0 z-50  mb-5 pt-4">
           <TierLimit />
         </div>
       </div>