+++ apps/platform/src/components/ui/input-otp.tsx
@@ -11,7 +11,7 @@ const InputOTP = React.forwardRef<
   <OTPInput
     className={cn('disabled:cursor-not-allowed', className)}
     containerClassName={cn(
-      'flex items-center justify-center gap-2 has-[:disabled]:opacity-50',
+      'flex items-center justify-center gap-2 has-disabled:opacity-50',
       containerClassName
     )}
     ref={ref}